srcs/backend/ai_services/ai_service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch
transformers
faster-whisper
requests
uvicorn
fastapi
tensorflow-cpu
numpy
opencv-python
python-multipart
<<<<<<< HEAD
=======
python-jose
python-dotenv
>>>>>>> 976ce7f9ad7db8f90acee698ae20c86670c5b353
